not getting data if metrics are disabled
